requirement.txt
django
djangorestframework
django-filter
request
django-environ